requirements.txt
pandas
streamlit
openai
pytorch